ERP/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bootstrap==3.3.7.1
psycopg2-binary==2.9.9
WTForms==3.0.1
python-dotenv==1.0.0